proj-power-rankings/src/requirements.txt
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
cloudevents==1.6.1
deprecation==2.1.0
Flask==2.2.2
functions-framework==3.2.0
google-auth==2.12.0
google-auth-oauthlib==0.5.3
gspread==5.6.0
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.1
packaging==21.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
urllib3==1.26.12
watchdog==2.1.9
Werkzeug==2.2.2
